fixed cross-screen touch bug
previously, when you continued touching the screen from the "you lost"
screen it would quickly go through the MainMenu way too fast and start
the game right away. since it was detected a touch press right after the
first frame of screen changes, it would act as if you started the game.

this was fixed keeping track if the player was touching before the "you
lost" and MainMenu screens first frames, and go to the next screen only
if the player wasn't previously touching.